% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Calculate Norm of Sparse Numeric Vector}
\usage{
norm(x)

\S4method{norm}{sparse_numeric}(x)
}
\arguments{
\item{x}{A sparse_numeric object}
}
\value{
A numeric scalar representing the Euclidean norm
}
\description{
Computes the squared norm of the vector, which is the square root of the sum of the squared individual elements of a vector.
}
\examples{
x <- as(c(3, 0, 4, 0, 0), "sparse_numeric")
norm(x)
# Result: sqrt(3^2 + 4^2) = 5

}
