% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot Two Sparse Numeric Vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional graphical parameters passed to plot()}
}
\value{
returns NULL, but creates a plot
}
\description{
Creates a scatter plot showing the overlapping non-zero elements of two sparse
numeric vectors. The x-axis represents position, and the y-axis represents value.
}
\examples{
x <- as(c(1, 0, 3, 4, 5, 0, 7), "sparse_numeric")
y <- as(c(2, 0, 3, 0, 6, 0, 8), "sparse_numeric")
plot(x, y)

}
