% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{sparse-arithmetic}
\alias{sparse-arithmetic}
\alias{+,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Arithmetic Operators for Sparse Numeric Vectors}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{A sparse_numeric object}

\item{e2}{A sparse_numeric object}
}
\value{
A sparse_numeric object
}
\description{
Arithmetic Operators for Sparse Numeric Vectors
}
