% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Add Two Sparse Numeric Vectors}
\usage{
sparse_add(x, y, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional arguments (currently unused)}
}
\value{
A sparse_numeric object containing the element-wise sum
}
\description{
Performs element-wise addition of two sparse numeric vectors of equal length.
}
\examples{
x <- as(c(1, 0, 3, 0, 5), "sparse_numeric")
y <- as(c(0, 2, 3, 0, 0), "sparse_numeric")
result <- sparse_add(x, y)
# Can also use: result <- x + y

}
