% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Compute Cross Product of Two Sparse Numeric Vectors}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional arguments (currently unused)}
}
\value{
A numeric scalar representing the cross product
}
\description{
Computes the cross product of two sparse numeric vectors of equal length.
}
\examples{
x <- as(c(1, 0, 3, 0, 5), "sparse_numeric")
y <- as(c(2, 0, 3, 0, 1), "sparse_numeric")
result <- sparse_crossprod(x, y)
# Result: 1*2 + 3*3 + 5*1 = 16

}
