% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiply Two Sparse Numeric Vectors Element-wise}
\usage{
sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional arguments (currently unused)}
}
\value{
A sparse_numeric object containing the element-wise product
}
\description{
Performs element-wise multiplication of two sparse numeric vectors of equal length.
}
\examples{
x <- as(c(2, 0, 3, 0, 5), "sparse_numeric")
y <- as(c(1, 2, 3, 0, 0), "sparse_numeric")
result <- sparse_mult(x, y)
# Can also use: result <- x * y

}
