% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\docType{methods}
\name{sparse_numeric-coercion}
\alias{sparse_numeric-coercion}
\alias{\S4method{coerce}{numeric,sparse_numeric}}
\alias{\S4method{coerce}{sparse_numeric,numeric}}
\title{Coercion Methods for sparse_numeric}
\arguments{
\item{from}{The object to be coerced.}
}
\description{
S4 methods used to convert between \code{numeric} and \code{sparse_numeric}.
}
