% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Calculate Mean of Sparse Numeric Vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments (currently unused)}
}
\value{
A numeric scalar representing the mean
}
\description{
Computes the arithmetic mean of a sparse numeric vector, accounting for all elements
including the implicit zeros.
}
\examples{
x <- as(c(2, 0, 6, 0, 0), "sparse_numeric")
mean(x)
# Result: (2 + 0 + 6 + 0 + 0) / 5 = 1.6

}
