% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\docType{class}
\name{sparse_numeric}
\alias{sparse_numeric}
\title{Sparse Numeric Class}
\description{
An S4 class that stores a sparse numeric vector using value/position
pairs.
}
\section{Slots}{

\describe{
\item{\code{value}}{A numeric vector of non-zero values.}

\item{\code{pos}}{An integer vector of positions corresponding to the values.}

\item{\code{length}}{An integer giving the total logical length of the vector.}
}}

\examples{
x <- new("sparse_numeric", value = c(1, 3), pos = c(1L, 4L), length = 5L)
y <- new("sparse_numeric", value = c(2), pos = c(4L), length = 5L)
x + y

}
