% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract Two Sparse Numeric Vectors}
\usage{
sparse_sub(x, y, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional arguments (currently unused)}
}
\value{
A sparse_numeric object containing the element-wise difference (x - y)
}
\description{
Performs element-wise subtraction of two sparse numeric vectors of equal length.
}
\examples{
x <- as(c(5, 0, 3, 0, 2), "sparse_numeric")
y <- as(c(1, 0, 3, 0, 0), "sparse_numeric")
result <- sparse_sub(x, y)
# Can also use: result <- x - y

}
