% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a Sparse Numeric Vector}
\usage{
standardize(x)

\S4method{standardize}{sparse_numeric}(x)
}
\arguments{
\item{x}{A sparse_numeric object}
}
\value{
A sparse_numeric object containing the standardized values
}
\description{
Standardizes a sparse numeric vector by subtracting the mean and dividing by the
standard deviation The result is returned as a sparse_numeric
object.
}
\examples{
x <- as(c(1, 0, 3, 0, 5, 0, 0), "sparse_numeric")
x_std <- standardize(x)
# Result has mean ≈ 0 and standard deviation = 1

}
