% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{*,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Multiply two sparse_numeric objects}
\usage{
\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1, e2}{sparse_numeric objects}
}
\value{
A sparse_numeric object
}
\description{
Multiply two sparse_numeric objects
}
